Moved the parser to another file
