+++ src/consensus.rs
@@ -14,12 +14,8 @@
     You should have received a copy of the GNU General Public License
     along with onyan.  If not, see <http://www.gnu.org/licenses/>.
 */
-use std::error::Error;
-use std::fs::File;
-use std::io::prelude::*;
-use std::io::BufReader;
 use std::io::BufRead;
-use nom::{anychar, IResult};
+
 /// Spec defined at https://gitweb.torproject.org/torspec.git/tree/dir-spec.txt
 /// Votes and consensuses are more strictly formatted than other documents
 /// in this specification, since different authorities must be able to
@@ -720,185 +716,3 @@ struct Router {}
 struct Threshold {}
 struct Param {}
 struct Authority {}
-
-/// The highest level object is a Document, which consists of one or more Items.
-#[derive(Debug)]
-pub struct Document {
-    items: Vec<Item>,
-}
-impl Document {
-    fn new(items: Vec<Item>) -> Self {
-        Document { items }
-    }
-}
-/// Document ::= (Item | NL)+
-#[cfg_attr(rustfmt, rustfmt_skip)]
-named!(
-    pub parse_document<Document>,
-    do_parse!(
-        items: many1!(
-            alt!(
-                do_parse!(
-                    item: item >>
-                    (Some(item))
-                ) |
-                do_parse!(
-                    newline >>
-                    (None)
-                )
-            )
-        ) >>
-        // Collect the option items into a vector to skip newlines
-        (Document::new(items.into_iter().flat_map(|optional_item| optional_item).collect()))
-    )
-);
-
-/// Every Item begins with a KeywordLine, followed by zero or more Objects.
-#[derive(Debug)]
-pub struct Item {
-    keyword_line: KeywordLine,
-    objects: Vec<Keyword>,
-}
-impl Item {
-    /// Constructor
-    // TODO: Add objects to the constructor
-    fn new(keyword_line: KeywordLine) -> Self {
-        Item {
-            keyword_line: keyword_line,
-            objects: vec![],
-        }
-    }
-}
-/// Item ::= KeywordLine Object*
-#[cfg_attr(rustfmt, rustfmt_skip)]
-named!(
-    item<Item>,
-    do_parse!(
-        keyword: keyword_line >>
-        objects: many0!(object) >>
-        (Item::new(keyword))
-    )
-);
-
-#[derive(Debug)]
-pub struct KeywordLine {
-    keyword: Keyword,
-    arguments: Option<Vec<char>>,
-}
-impl KeywordLine {
-    /// Constructor
-    fn new(keyword: Keyword, arguments: Option<Vec<char>>) -> Self {
-        KeywordLine { keyword, arguments }
-    }
-}
-/// KeywordLine ::= Keyword NL | Keyword WS ArgumentChar+ NL
-#[cfg_attr(rustfmt, rustfmt_skip)]
-named!(
-    keyword_line<KeywordLine>,
-    alt_complete!(
-        do_parse!(
-            keyword: keyword >>
-            newline >>
-            (KeywordLine::new(keyword, None))
-        ) |
-        do_parse!(
-            keyword: keyword >>
-            whitespace >>
-            arguments: many1!(argument_char) >>
-            newline >>
-            //TODO: Add arguments back
-            (KeywordLine::new(keyword, Some(arguments))) 
-        )
-    )
-);
-
-/// Keyword = KeywordChar+
-/// KeywordChar ::= 'A' ... 'Z' | 'a' ... 'z' | '0' ... '9' | '-'
-#[cfg_attr(rustfmt, rustfmt_skip)]
-named!(
-    keyword<Keyword>,
-    do_parse!(
-        keyword_chars: many1!(
-            alt!(
-                upper_alphabet |
-                lower_alphabet |
-                number |
-                char!('-')
-            )
-        ) >> (Keyword::from_chars(keyword_chars))
-    )
-);
-#[derive(Debug)]
-pub struct Keyword {
-    keyword: String,
-}
-impl Keyword {
-    /// Constructor
-    fn from_chars(characters: Vec<char>) -> Self {
-        Keyword {
-            keyword: characters.into_iter().collect(),
-        }
-    }
-}
-/// Lowercase alphabet
-named!(lower_alphabet<char>, one_of!("abcdefghijklmnopqrstuvwxyz"));
-/// Uppercase alphabet
-named!(upper_alphabet<char>, one_of!("ABCDEFGHIJKLMNOPQRSTUVWXYZ"));
-/// Numbers
-named!(number<char>, one_of!("0123456789"));
-
-/// ArgumentChar ::= any printing ASCII character except NL.
-// TODO: generate this in a better way
-named!(argument_char<char>, one_of!(" !\"#$%&'()*+,-./0123456789:;<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\\]^_`abcdefghijklmnopqrstuvwxyz{|}~"));
-
-/// Object ::= BeginLine Base64-encoded-data EndLine
-#[cfg_attr(rustfmt, rustfmt_skip)]
-named!(
-    object<Vec<char>>,
-    do_parse!(
-        begin_line: begin_line >>
-        data: many1!(
-            one_of!("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789+/=\n")
-        ) >> 
-        end_line: end_line >>
-        (data)
-    )
-);
-/// BeginLine ::= "-----BEGIN " Keyword "-----" NL
-#[cfg_attr(rustfmt, rustfmt_skip)]
-named!(
-    begin_line<Keyword>,
-    delimited!(
-        tag!("-----BEGIN "),
-        keyword,
-        do_parse!(tag!("-----") >> newline >> ())
-    )
-);
-/// EndLine ::= "-----END " Keyword "-----" NL
-#[cfg_attr(rustfmt, rustfmt_skip)]
-named!(
-    end_line<Keyword>,
-    delimited!(
-        tag!("-----END "),
-        keyword,
-        do_parse!(tag!("-----") >> newline >> ())
-    )
-);
-/// NL = The ascii LF character (hex value 0x0a).
-named!(newline<char>, char!('\n'));
-/// WS = (SP | TAB)+
-#[cfg_attr(rustfmt, rustfmt_skip)]
-named!(
-    whitespace<Vec<char>>,
-    many1!(one_of!(" \t"))
-);
-
-/// Test
-#[test]
-fn test_parse_document() {
-    let mut file = File::open("test/barebones.consensus").expect("file not found");
-    let mut file_bytes: Vec<u8> = Vec::new();
-    file.read_to_end(&mut file_bytes).unwrap();
-    let a = parse_document(&file_bytes);
-    println!("{:#?}", a);
-}
